backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
brotli==1.2.0
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cssselect2==0.9.0
Django==6.0.2
django-cors-headers==4.9.0
django-filter==25.1
django-redis==6.0.0
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fonttools==4.61.1
gunicorn==25.1.0
idna==3.11
kombu==5.6.2
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==3.0
pydyf==0.12.1
PyJWT==2.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.2.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
tinycss2==1.5.1
tinyhtml5==2.0.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
weasyprint==68.1
webencodings==0.5.1
zopfli==0.4.1
